moving convolve tests to own file
